# docs/swagger.yaml
basePath: /
definitions:
...
  addTask.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  addTask.Tasks:
    properties:
      date:
        type: string
      description:
        type: string
      id:
        type: string
      pk:
        type: string
      size:
        type: number
      sk:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      username:
        type: string
    type: object
  addTask.UserRequest:
    properties:
      description:
        type: string
      size:
        type: number
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
...
  changeStatus.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  changeStatus.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  changeStatus.UserRequest:
    properties:
      id:
        type: string
      token:
        type: string
    type: object
...
  deleteTask.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  deleteTask.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  deleteTask.UserRequest:
    properties:
      id:
        type: string
      token:
        type: string
    type: object
...
  getTag.Ebook:
    type: object
  getTag.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  getTag.Files:
    type: object
  getTag.Passwords:
    type: object
  getTag.SuccessResponse:
    properties:
      ebooks:
        items:
          $ref: '#/definitions/getTag.Ebook'
        type: array
      files:
        items:
          $ref: '#/definitions/getTag.Files'
        type: array
      passwords:
        items:
          $ref: '#/definitions/getTag.Passwords'
        type: array
      tasks:
        items:
          $ref: '#/definitions/getTag.Tasks'
        type: array
    type: object
  getTag.Tasks:
    type: object
...
  getTask.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  getTask.Tasks:
    properties:
      date:
        type: string
      description:
        type: string
      id:
        type: string
      pk:
        type: string
      size:
        type: number
      sk:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      username:
        type: string
    type: object
  getTask.UserRequest:
    properties:
      id:
        type: string
      token:
        type: string
    type: object
...
  listTasks.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  listTasks.Tasks:
    properties:
      description:
        type: string
      id:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  listTasks.TasksList:
    properties:
      tasks:
        items:
          $ref: '#/definitions/listTasks.Tasks'
        type: array
    type: object
  listTasks.UserRequest:
    properties:
      token:
        type: string
    type: object
  updateAvatarUrl.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  updateAvatarUrl.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  updateAvatarUrl.UserRequest:
    properties:
      avatar_url:
        type: string
    type: object
...
  updateTypeVerification.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  updateTypeVerification.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  updateTypeVerification.UserRequest:
    properties:
      type:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for personal server.
  title: grp@server
  version: "1.0"
paths:
...
  /tags/:tag:
    get:
      consumes:
      - application/json
      description: Get all domain with relationship with the user tags from a user
        with the header token
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tag name
        in: path
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/getTag.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/getTag.ErrorResponse'
      summary: Get all domains by tags from a user
      tags:
      - tags
  /tasks:
    get:
      consumes:
      - application/json
      description: Get all tasks from a user with token
      parameters:
      - description: Token in headers
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/listTasks.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/listTasks.TasksList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/listTasks.ErrorResponse'
      summary: Get all tasks from a user
      tags:
      - task
    post:
      consumes:
      - application/json
      description: Add a task with the request body
      parameters:
      - description: Request body
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/addTask.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/addTask.Tasks'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/addTask.ErrorResponse'
      summary: Add a task
      tags:
      - task
  /tasks/:id:
    delete:
      consumes:
      - application/json
      description: Delete a task with the url id and token
      parameters:
      - description: url id and token
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/deleteTask.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/deleteTask.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/deleteTask.ErrorResponse'
      summary: Delete a task
      tags:
      - task
    get:
      consumes:
      - application/json
      description: Get a task with the url id and token
      parameters:
      - description: url id and token
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/getTask.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/getTask.Tasks'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/getTask.ErrorResponse'
      summary: Get a task
      tags:
      - task
    patch:
      consumes:
      - application/json
      description: Change status task with the request body and url id and token
      parameters:
      - description: url id and token
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/changeStatus.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/changeStatus.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/changeStatus.ErrorResponse'
      summary: Change status task
      tags:
      - task
...
  /users/avatar_url:
    patch:
      consumes:
      - application/json
      description: Update a user avatar with the request body
      parameters:
      - description: avatar url in body and token in headers
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/updateAvatarUrl.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/updateAvatarUrl.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/updateAvatarUrl.ErrorResponse'
      summary: Update a user avatar
      tags:
      - user
...
  /users/type:
    patch:
      consumes:
      - application/json
      description: Update a user Type verification with the request body
      parameters:
      - description: Type verification in body and token in headers
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/updateTypeVerification.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/updateTypeVerification.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/updateTypeVerification.ErrorResponse'
      summary: Update a user Type verification
      tags:
      - user
swagger: "2.0"